contracts continued
