+++ examples/ink_e2e/interact.ts
@@ -1,0 +1,5 @@
+import { address } from "./deploy.ts"
+
+const instance = ink.InkContractRune.fromAddress(address)
+
+instance.call(...)
