+++ patterns/ink/known.ts
