+++ patterns/ink/mod.ts
@@ -1,12 +1,9 @@
 // moderate
 
-export * from "./codecs.ts"
-export * from "./ContractError.ts"
-export * from "./ContractInstanceRune.ts"
-export * from "./ContractInstantiationExtrinsicRune.ts"
-export * from "./ContractInstantiationExtrinsicStatusRune.ts"
-export * from "./ContractInstantiationSignedExtrinsicRune.ts"
-export * from "./ContractMetadata.ts"
-export * from "./ContractMsgRune.ts"
-export * from "./ContractMsgSignedRune.ts"
-export * from "./ContractRune.ts"
+export * from "./InkMetadata.test.ts"
+export * from "./InkMetadata.ts"
+export * from "./InkMetadataRune.ts"
+export * from "./InkMsgRune.ts"
+export * from "./InkRune.ts"
+export * from "./known.ts"
+export * from "./util/mod.ts"
